feat(grid): dump raw rows and draw every parasite shape in the targeted run
grid/target_parasite.py kept rows only in memory and wrote just the aggregate,
so the raw per-reading data behind target_parasite.json was never archived.
--dump PATH writes every raw row (JSONL, one reading per line, each carrying
its own ts). Verified byte-identical: a replay with --dump alone reproduces
grid/results/target_parasite.json exactly (git status clean after the run).

The shape draw was seed % 3 into a sorted SHAPES list, never looking at the
cell. With the 3 seeds this generator uses (11, 23, 37), two of them collide
on the same index and the first sorted shape (fold) never ran, at any cell,
for any variant, in any prior run including the published one. --all-shapes
draws on (seed + cell index) % 3 instead, which the archival run (2,592
readings, shape counts 648/648/648 across fold, speck and stroke) confirms
covers all three shapes evenly.

Running the archival pass surfaced a real, pre-existing defect this task's
scope covers fixing: preflight/parasite.py's speck() computes its Gaussian
blur radius with np.sqrt, and Pillow 12.3.0 (the fresh venv built for this
study) compares GaussianBlur's radius against the tuple (0, 0) before its
scalar check, which raises on a numpy scalar. A plain float() around the
radius fixes it; confirmed to reproduce on the unmodified speck() too, so
this is unrelated to the flags above and was blocking the archival run
regardless of them.

Full suite green before and after (72 passed, including the 4 new tests on
shape_for()).

+++ preflight/parasite.py
@@ -39,7 +39,10 @@ def speck(grey, zone, dpi, fraction, rng, darkness=45):
     im = Image.fromarray(grey)
     d = ImageDraw.Draw(im)
     d.ellipse([cx - rx, cy - ry, cx + rx, cy + ry], fill=int(darkness))
-    return np.asarray(im.filter(ImageFilter.GaussianBlur(max(0.6, ry * 0.12))))
+    # float(): ry is a numpy scalar (np.sqrt), and Pillow's GaussianBlur.filter() compares its
+    # radius against the tuple (0, 0) before the scalar check; a numpy scalar there broadcasts
+    # into a 2-element boolean array and "if" on it raises ValueError. A plain float sidesteps it.
+    return np.asarray(im.filter(ImageFilter.GaussianBlur(float(max(0.6, ry * 0.12)))))
 
 
 def fold(grey, zone, dpi, fraction, rng):
